+++ src/Exception/FailedToCloseSession.php
@@ -1,8 +1,0 @@
-<?php
-declare(strict_types = 1);
-
-namespace Innmind\HttpSession\Exception;
-
-final class FailedToCloseSession extends RuntimeException
-{
-}