+++ src/Manager.php
@@ -4,12 +4,25 @@ declare(strict_types = 1);
 namespace Innmind\HttpSession;
 
 use Innmind\Http\Message\ServerRequest;
+use Innmind\Immutable\{
+    Maybe,
+    SideEffect,
+};
 
 interface Manager
 {
-    public function start(ServerRequest $request): Session;
-    public function get(ServerRequest $request): Session;
-    public function contains(ServerRequest $request): bool;
-    public function save(ServerRequest $request): void;
-    public function close(ServerRequest $request): void;
+    /**
+     * @return Maybe<Session>
+     */
+    public function start(ServerRequest $request): Maybe;
+
+    /**
+     * @return Maybe<SideEffect>
+     */
+    public function save(Session $session): Maybe;
+
+    /**
+     * @return Maybe<SideEffect>
+     */
+    public function close(Session $session): Maybe;
 }